requirements.txt
google_api_python_client==1.12.8
google_auth_oauthlib==1.0.0
protobuf==3.20.3
requests==2.27.1